+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Bootstrap installation using the `init` command
+
 ## [0.1.0] - 2025-04-02
 
 ### Added